+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add `AnyEraScript`
 * Add `AnyEraSpendingPurpose`, `AnyEraMintingPurpose`, `AnyEraCertifyingPurpose`, `AnyEraRewardingPurpose`, `AnyEraVotingPurpose`, `AnyEraProposingPurpose`, `AnyEraGuardingPurpose` patterns
 * Re-export `DijkstraEraScript`, `toGuardingPurpose` and `GuardingPurpose` pattern
+* Add `producedTxOuts` and `mkCollateralTxIn`
 * Add `AnyEraTx`
 * Add `AnyEraTxBody`
 * Add `ttlToValidityInterval`